requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg[binary]>=3.1
psycopg-pool>=3.2
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
